requirements.txt
jupyterlab==4.2.5
python-dotenv==1.0.1
langchain-community==0.3.29
langchain-openai==0.3.32
langchain-huggingface==0.3.1
sentence-transformers==5.1.0
langchain-groq==0.3.8
langchainhub==0.1.21
pypdf==6.0.0
faiss-cpu==1.12.0
plotly==6.3.0
pandas==2.3.2
streamlit==1.41.1
